refactor: unentangle experiment builder from engine
Here the idea is that the logic for creating an experiment lives in
internal/registry along with a list of experiments.

In turn, this design will allow us soon to make it very easy to add a
new experiment, by making each experiment register itself.

However, before doing that, I need to refactor the code so that
each experiment and the engine can both import from registry.

    +---------------+                 +--------+
    | An experiment |                _| engine |
    +---------------+               | +--------+
              |                     |
              |                     |
              |      +----------+   |
              `----->| registry |<--.  imports
          imports    +----------+

Once the above is possible, then each experiment can register itself.
