+++ internal/registry/doc.go
@@ -1,0 +1,2 @@
+// Package registry contains a registry of all the available experiments.
+package registry
